pyproject.toml
[project]
name = "social-media-mongodb-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Aarya Ranjit",email = "aaryaranjit03@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "motor (>=3.7.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
